fix(sampling) Use Discover entity instead of transactions
There were a couple issues here: the first is that the get_tag_keys only ever
uses the events or discover dataset. The second was that the callback metrics
were being called for every referrer.
